# backend/docs/swagger.yaml
basePath: /
definitions:
...
  router.authProviderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.authProviderResponse'
        type: array
    type: object
...
  router.authProviderResponse:
    properties:
      config:
        type: object
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  router.dhcpExclusionResponse:
    properties:
      endIp:
        type: string
      externalId:
        type: string
      id:
        type: string
      lastSyncedAt:
        type: string
      scopeId:
        type: string
      startIp:
        type: string
    type: object
...
  router.errorDocResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  router.notificationChannelListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.notificationChannelResponse'
        type: array
      total:
        type: integer
    type: object
...
  router.notificationChannelResponse:
    properties:
      config:
        type: object
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      passwordResetEnabled:
        type: boolean
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: ZoneLease Windows DNS / DHCP 统一管理控制台后端 API，提供认证、服务器登记、DNS/DHCP 资源管理、刷新事件和审计查询接口。
  title: ZoneLease API
  version: "1.0"
paths:
...
  /api/dhcp/exclusions:
    post:
      consumes:
      - application/json
      description: 转发到 DHCP Agent 创建 Windows DHCP 排除范围，成功后按当前作用域延迟合并创建 runtime.refresh.dhcp.scope
        局部刷新任务。
      parameters:
      - description: DHCP 排除范围参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.dhcpExclusionResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.dhcpExclusionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 创建 DHCP 排除范围
      tags:
      - DHCP
...
  /api/settings/auth-providers:
    get:
      description: 当前支持 AD/LDAP 目录认证配置，启用后登录页显示对应认证方式。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.authProviderListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 获取认证配置列表
      tags:
      - Settings
...
  /api/settings/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.notificationChannelListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 获取通知媒介配置
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"